Emakefile, systools rel file, added to Makefile
I'm going to switch to reltool from systools and I'm going to switch to
/rel from /ebin for release files, but I want to check things in now for
safe keeping. The compile target in Makefile is redundant, but that's
okay; I was reminded of some things when I wrote it.

www.learnyousomeerlang.com was a tremendous resource when working on
this. All the documentation is available for Erlang, but it doesn't
cover the most up-to-date, standard, or most popular practices. I spent
some time using systools before I read (again) about reltool on LYSE.